-Adds BitmapCompressor class -Separates getAgendaItems and updateAgendaItemCache
